V2
Now invoice number can be read from a text file.

The text file is read using Javascript. On clicking exit, an updates txt file is served to download, as there is no way to use JS to update a text file. Moreover the aim was not to use any server side language.